+++ tools/contributed/traas/src/main/java/de/tudresden/sumo/cmd/Vehicle.java
@@ -1,19 +1,22 @@
 /****************************************************************************/
 // Eclipse SUMO, Simulation of Urban MObility; see https://eclipse.org/sumo
-// Copyright (C) 2017-2019 German Aerospace Center (DLR) and others.
+// Copyright (C) 2017-2020 German Aerospace Center (DLR) and others.
 // TraaS module
 // Copyright (C) 2016-2017 Dresden University of Technology
-// This program and the accompanying materials
-// are made available under the terms of the Eclipse Public License v2.0
-// which accompanies this distribution, and is available at
-// http://www.eclipse.org/legal/epl-v20.html
-// SPDX-License-Identifier: EPL-2.0
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License 2.0 which is available at
+// https://www.eclipse.org/legal/epl-2.0/
+// This Source Code may also be made available under the following Secondary
+// Licenses when the conditions for such availability set forth in the Eclipse
+// Public License 2.0 are satisfied: GNU General Public License, version 2
+// or later which is available at
+// https://www.gnu.org/licenses/old-licenses/gpl-2.0-standalone.html
+// SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-or-later
 /****************************************************************************/
 /// @file    Vehicle.java
 /// @author  Mario Krumnow
 /// @author  Evamarie Wiessner
 /// @date    2016
-/// @version $Id$
 ///
 //
 /****************************************************************************/
@@ -322,7 +325,7 @@ public class Vehicle {
      * @param edgeID edge id
      * @return the effort
      */
-    public static SumoCommand getEffort(String vehID, int time, String edgeID) {
+    public static SumoCommand getEffort(String vehID, double time, String edgeID) {
         Object[] array = new Object[] {time, edgeID};
         return new SumoCommand(Constants.CMD_GET_VEHICLE_VARIABLE, Constants.VAR_EDGE_EFFORT, vehID, array, Constants.RESPONSE_GET_VEHICLE_VARIABLE, Constants.TYPE_DOUBLE);
     }
